Extra flag mode bug has fixed
In the previous versions the extra flag calculation was based on a bare comparison, that only worked for numbers. Strings and other classes wont trigger the extra flag insertion. Therefore the insertion was moved into the actual key replacement function. When a key was parsed, the corresponding flag will be inserted as well. No comparison will happen.